+++ src/features/wallet/screens/TransferDetails/form.ts
@@ -7,14 +7,15 @@ export const getTransferFormInitialValues = (): TransferFormValues => ({
   amount: '',
 });
 
-export const useTransferFormValidation = () =>
+export const useTransferFormValidation = (maxAmount: number) =>
   useMemo(
     () =>
       object().shape({
         amount: number()
           .typeError('Amount must be a number')
-          .min(0, 'Amount must be positive')
+          .notOneOf([0], 'Amount cannot be 0')
+          .max(maxAmount, `Amount must not exceed ${maxAmount}`)
           .required('Amount is required'),
       }),
-    [],
+    [maxAmount],
   );